+++ src/ResourceManager/AzureBatch/ChangeLog.md
@@ -18,12 +18,12 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Release new cmdlet Get-AzureBatchPoolNodeCounts
+* Release new cmdlet Start-AzureBatchComputeNodeServiceLogUpload
 
 ## Version 4.0.7
 * Set minimum dependency of module to PowerShell 5.0
 * Updated New-AzureBatchPool documentation to remove deprecated example
-* Release new cmdlet Get-AzureBatchPoolNodeCounts
-* Release new cmdlet Start-AzureBatchComputeNodeServiceLogUpload
 
 ## Version 4.0.6
 * Updated to the latest version of the Azure ClientRuntime